requirements.txt
google-genai>=0.5.0
ddgs>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
